2002-07-19  JP Rosevear  <jpr@ximian.com>
	* libpisock/unixserial.c: Update for LOG changes

	* libpisock/sys.c: ditto

	* libpisock/socket.c: ditto

	* libpisock/slp.c: ditto

	* libpisock/pi-file.c: ditto

	* libpisock/padp.c: ditto

	* libpisock/net.c: ditto

	* libpisock/linuxusb.c: ditto

	* libpisock/inet.c: ditto

	* libpisock/freebsdusb.c: ditto

	* libpisock/dlp.c: ditto

	* libpisock/cmp.c: ditto

	* tests/dlp-test.c: change CHECK to CHECK_RESULT so we don't
	conflict with the debug macro, reflect LOG define changes

	* include/pi-debug.h: redefine LOG so it is not a variadic macro

+++ libpisock/palmpix.c
@@ -1393,7 +1393,7 @@ int unpack_PalmPix (struct PalmPixState *s,
 	for (k = 0; k < 4; k++) 
 	  {
 	     unsigned long nbytes, offset;
-	     int16_t lastWord = 0;
+	     int32_t lastWord = 0;
 	     int j;
 	     
 	     for (nbytes = 0; nbytes < h->chansize[k]; recno++) 
@@ -1419,7 +1419,7 @@ int unpack_PalmPix (struct PalmPixState *s,
 	     
 	     for (j = 1; j < rawHeight; j++) 
 	       {
-		  uint16_t eaten;
+		  uint32_t eaten;
 		  DecodeRow (&raw[offset],
 			     &chan[k][rawWidth * (j - 1)],
 			     &chan[k][rawWidth * j],
